Add files via upload
Reflection:
This design thinking enables me to think critically about the problem and find a solution that fits the user's needs. Not only that, design thinking also allows me to work in a team and build my confidence in team management. I am grateful for leading this project and successfully creating a working prototype for me and my team.